JS: fix help anim for older WebGL. Fix fire anim.
The animated dash lines in the help boxes were animated as 20 full
size help boxes which meant the width of the spritesheet was well over
4096 pixels which breaks on some older WebGL versions. So the layout
now uses a grid of animations (instead of the horizontal line) and the
dash animations are rendered at half size and scaled up (since it
looks about the same).

Also, add winning score divisor to the HUD stats.